remove unused validateTopicExists in cli package as it has been moved to cli/deploy package
