+++ internal/pkg/cli/deploy/deploy_test.go
@@ -949,3 +949,59 @@ func TestSvcDeployOpts_stackConfiguration_worker(t *testing.T) {
 		})
 	}
 }
+
+func Test_validateTopicsExist(t *testing.T) {
+	mockApp := "app"
+	mockEnv := "env"
+	mockAllowedTopics := []string{
+		"arn:aws:sqs:us-west-2:123456789012:app-env-database-events",
+		"arn:aws:sqs:us-west-2:123456789012:app-env-database-orders",
+		"arn:aws:sqs:us-west-2:123456789012:app-env-api-events",
+	}
+	duration10Hours := 10 * time.Hour
+	testGoodTopics := []manifest.TopicSubscription{
+		{
+			Name:    aws.String("events"),
+			Service: aws.String("database"),
+		},
+		{
+			Name:    aws.String("orders"),
+			Service: aws.String("database"),
+			Queue: manifest.SQSQueueOrBool{
+				Advanced: manifest.SQSQueue{
+					Retention: &duration10Hours,
+				},
+			},
+		},
+	}
+	testCases := map[string]struct {
+		inTopics    []manifest.TopicSubscription
+		inTopicARNs []string
+
+		wantErr string
+	}{
+		"empty subscriptions": {
+			inTopics:    nil,
+			inTopicARNs: mockAllowedTopics,
+		},
+		"topics are valid": {
+			inTopics:    testGoodTopics,
+			inTopicARNs: mockAllowedTopics,
+		},
+		"topic is invalid": {
+			inTopics:    testGoodTopics,
+			inTopicARNs: []string{},
+			wantErr:     "SNS topic app-env-database-events does not exist in environment env",
+		},
+	}
+	for name, tc := range testCases {
+		t.Run(name, func(t *testing.T) {
+			err := validateTopicsExist(tc.inTopics, tc.inTopicARNs, mockApp, mockEnv)
+			if tc.wantErr != "" {
+				require.EqualError(t, err, tc.wantErr)
+			} else {
+				require.NoError(t, err)
+			}
+		})
+	}
+}
