+++ internal/pkg/cli/validate_test.go
@@ -8,11 +8,9 @@ import (
 	"fmt"
 	"strings"
 	"testing"
-	"time"
 
 	"github.com/aws/copilot-cli/internal/pkg/workspace"
 
-	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/copilot-cli/internal/pkg/manifest"
 
 	"github.com/spf13/afero"
@@ -883,62 +881,6 @@ func Test_validateSubscribe(t *testing.T) {
 				require.NoError(t, err)
 			} else {
 				require.EqualError(t, err, tc.wantErr.Error())
-			}
-		})
-	}
-}
-
-func Test_validateTopicsExist(t *testing.T) {
-	mockApp := "app"
-	mockEnv := "env"
-	mockAllowedTopics := []string{
-		"arn:aws:sqs:us-west-2:123456789012:app-env-database-events",
-		"arn:aws:sqs:us-west-2:123456789012:app-env-database-orders",
-		"arn:aws:sqs:us-west-2:123456789012:app-env-api-events",
-	}
-	duration10Hours := 10 * time.Hour
-	testGoodTopics := []manifest.TopicSubscription{
-		{
-			Name:    aws.String("events"),
-			Service: aws.String("database"),
-		},
-		{
-			Name:    aws.String("orders"),
-			Service: aws.String("database"),
-			Queue: manifest.SQSQueueOrBool{
-				Advanced: manifest.SQSQueue{
-					Retention: &duration10Hours,
-				},
-			},
-		},
-	}
-	testCases := map[string]struct {
-		inTopics    []manifest.TopicSubscription
-		inTopicARNs []string
-
-		wantErr string
-	}{
-		"empty subscriptions": {
-			inTopics:    nil,
-			inTopicARNs: mockAllowedTopics,
-		},
-		"topics are valid": {
-			inTopics:    testGoodTopics,
-			inTopicARNs: mockAllowedTopics,
-		},
-		"topic is invalid": {
-			inTopics:    testGoodTopics,
-			inTopicARNs: []string{},
-			wantErr:     "SNS topic app-env-database-events does not exist in environment env",
-		},
-	}
-	for name, tc := range testCases {
-		t.Run(name, func(t *testing.T) {
-			err := validateTopicsExist(tc.inTopics, tc.inTopicARNs, mockApp, mockEnv)
-			if tc.wantErr != "" {
-				require.EqualError(t, err, tc.wantErr)
-			} else {
-				require.NoError(t, err)
 			}
 		})
 	}
